Some refactored app and added some functionality.

+++ src/main/java/com/gmail/voronovskyi/yaroslav/telegramjokebot/repository/IJokeRepository.java
@@ -3,5 +3,5 @@ package com.gmail.voronovskyi.yaroslav.telegramjokebot.repository;
 import com.gmail.voronovskyi.yaroslav.telegramjokebot.model.Joke;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IJokeRepository extends JpaRepository<Joke, Long> {
+public interface IJokeRepository extends JpaRepository<Joke, Integer> {
 }
